Refactored dfs_schema generation code
